+++ cli/build.ts
@@ -13,9 +13,14 @@ interface Options {
   config?: string;
   serve?: boolean;
   watch?: boolean;
+  dev?: boolean;
 }
 
-export default function ({ config, serve, watch }: Options) {
+export default function ({ config, serve, watch, dev }: Options) {
+  if (dev) {
+    Deno.env.set("LUME_ENV", "development");
+  }
+
   return build(config, serve, watch);
 }