+++ mod.ts
@@ -33,6 +33,12 @@ export default function lume(
     options = merge(options, getOptionsFromCli());
   }
 
+  if (options.dev) {
+    Deno.env.set("LUME_ENV", "development");
+  } else {
+    Deno.env.delete("LUME_ENV");
+  }
+
   const site = new Site(options as Partial<SiteOptions>);
 
   // Ignore the .git folder and .DS_Store (macOS) files by the watcher
@@ -83,8 +89,8 @@ function getOptionsFromCli(): DeepPartial<SiteOptions> {
     overrides.quiet = options.quiet;
   }
 
-  if (options.dev) {
-    overrides.dev = options.dev;
+  if (Deno.env.get("LUME_ENV") === "development") {
+    overrides.dev = true;
   }
 
   if (options.port) {